+++ .github/workflows/release.yaml
@@ -10,7 +10,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        image: ["dotbot.gateway", "dotbot.migration"]
+        image: ["dotbot.api", "dotbot.migration"]
     permissions:
       contents: write
       packages: write
