+++ Dotbot.sln
@@ -2,13 +2,13 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{CC37BA52-7578-490B-94DA-79A705D9B223}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dotbot.Gateway", "src\Dotbot.Gateway\Dotbot.Gateway.csproj", "{ABBF52EF-E4E2-4662-A915-6348F5DDC083}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dotbot.Api", "src\Dotbot.Api\Dotbot.Api.csproj", "{ABBF52EF-E4E2-4662-A915-6348F5DDC083}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceDefaults", "src\ServiceDefaults\ServiceDefaults.csproj", "{3245799D-63FF-48FD-BE1D-183F3F05A802}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{8378DB18-9725-4D55-BA59-7EED997FBC0B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dotbot.Gateway.UnitTests", "tests\Dotbot.Gateway.UnitTests\Dotbot.Gateway.UnitTests.csproj", "{3DB9E98D-953A-44D7-A32A-2B8FB76AEFDA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dotbot.Api.UnitTests", "tests\Dotbot.Api.UnitTests\Dotbot.Api.UnitTests.csproj", "{3DB9E98D-953A-44D7-A32A-2B8FB76AEFDA}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dotbot.Infrastructure", "src\Dotbot.Infrastructure\Dotbot.Infrastructure.csproj", "{75424D92-82F1-4450-B339-5D24B6ED2C99}"
 EndProject
