+++ src/Dotbot.Api/Dockerfile
@@ -1,4 +1,4 @@
-# renovate: datasource=dotnet-version depName=dotnet-runtime
+﻿# renovate: datasource=dotnet-version depName=dotnet-runtime
 ARG DOTNET_RUNTIME=mcr.microsoft.com/dotnet/aspnet:9.0.1-bookworm-slim
 # renovate: datasource=dotnet-version depName=dotnet-sdk
 ARG DOTNET_SDK=mcr.microsoft.com/dotnet/sdk:9.0.102
@@ -6,7 +6,7 @@ ARG DOTNET_SDK=mcr.microsoft.com/dotnet/sdk:9.0.102
 FROM ${DOTNET_RUNTIME} AS base
 WORKDIR /app
 EXPOSE 8080
-USER 1024
+USER 2048
 
 FROM ${DOTNET_SDK} AS build
 ARG BUILD_CONFIGURATION=Release
@@ -14,7 +14,6 @@ WORKDIR /src
 COPY ["nuget.config", "."]
 COPY ["Directory.Build.props", "."]
 COPY ["Directory.Packages.props", "."]
-COPY ["Dotbot.sln", "."]
 COPY ["src/Dotbot.Api/Dotbot.Api.csproj", "Dotbot.Api/Dotbot.Api.csproj"]
 COPY ["src/Dotbot.Infrastructure/Dotbot.Infrastructure.csproj", "Dotbot.Infrastructure/Dotbot.Infrastructure.csproj"]
 COPY ["src/ServiceDefaults/ServiceDefaults.csproj", "ServiceDefaults/ServiceDefaults.csproj"]
@@ -22,17 +21,15 @@ COPY ["src/ServiceDefaults/ServiceDefaults.csproj", "ServiceDefaults/ServiceDefa
 COPY ["src/Dotbot.Api/", "Dotbot.Api"]
 COPY ["src/Dotbot.Infrastructure/", "Dotbot.Infrastructure"]
 COPY ["src/ServiceDefaults/", "ServiceDefaults"]
+RUN dotnet restore "Dotbot.Api/Dotbot.Api.csproj"
+RUN dotnet build "Dotbot.Api/Dotbot.Api.csproj" -c $BUILD_CONFIGURATION -o /app/build
 
-# Build the migrationbundle here
-FROM build as migrationbuilder
-ENV PATH $PATH:/root/.dotnet/tools
-RUN dotnet tool install --global dotnet-ef
-RUN mkdir /migrations
-RUN dotnet ef migrations bundle -s /src/Dotbot.Api -p /src/Dotbot.Infrastructure -c DotbotContext --self-contained -r linux-x64 -o /migrations/migration
+FROM build AS publish
+ARG BUILD_CONFIGURATION=Release
+RUN dotnet publish "Dotbot.Api/Dotbot.Api.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
 
-FROM ${DOTNET_RUNTIME} as initcontainer
-ENV CONNECTIONSTRING=""
-COPY --from=migrationbuilder /migrations /migrations
-RUN chmod 755 /migrations/migration
-WORKDIR /migrations
-ENTRYPOINT ./migration --connection "$CONNECTIONSTRING"
+FROM base AS final
+WORKDIR /app
+ENV ASPNETCORE_URLS=http://+:8080
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "Dotbot.Api.dll"]
