+++ src/Dotbot.Api/Dto/XkcdComic.cs
@@ -1,0 +1,10 @@
+﻿namespace Dotbot.Api.Dto;
+
+public class XkcdComic
+{
+    public int ComicNumber { get; set; }
+    public DateTime DatePosted { get; set; }
+    public string AltText { get; set; } = null!;
+    public string ImageUrl { get; set; } = null!;
+    public string Title { get; set; } = null!;
+}
