+++ src/Dotbot.Api/Readme.md
@@ -1,0 +1,2 @@
+# Dotbot.Api
+This web api uses Discord's interaction API functionality to drive a discord bot that uses slash commands
