+++ src/Dotbot.Api/Settings/Discord.cs
@@ -1,0 +1,7 @@
+namespace Dotbot.Api.Settings;
+
+public class Discord
+{
+    public string Token { get; init; } = null!;
+    public string BucketEnvPrefix { get; set; } = null!;
+}
