--- src/Dotbot.Gateway/Application/Queries/IGuildQueries.cs
@@ -1,9 +1,0 @@
-using Dotbot.Infrastructure.Entities;
-
-namespace Dotbot.Gateway.Application.Queries;
-
-public interface IGuildQueries
-{
-    Task<IEnumerable<CustomCommand>> GetAllCustomCommands(string externalId);
-    Task<IEnumerable<CustomCommand>> GetCustomCommandsByFuzzySearchOnNameAsync(string externalId, string name);
-}
