--- src/Dotbot.Gateway/Discord.http
@@ -1,6 +1,0 @@
-@Discord_HostAddress = http://localhost:5292
-
-GET {{Discord_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
