--- src/Dotbot.Gateway/Dto/FileDetails.cs
@@ -1,15 +1,0 @@
-namespace Dotbot.Gateway.Dto;
-
-public class FileDetails
-{
-    public FileDetails(Stream fileContent, string filename, string parentName)
-    {
-        FileContent = fileContent;
-        Filename = filename;
-        ParentName = parentName;
-    }
-
-    public Stream FileContent { get; set; }
-    public string Filename { get; set; }
-    public string ParentName { get; set; }
-}
