--- src/Dotbot.Gateway/Program.Testing.cs
@@ -1,3 +1,0 @@
-namespace Dotbot.Gateway;
-
-public partial class Program { }
