--- src/Dotbot.Gateway/Readme.md
@@ -1,2 +1,0 @@
-# Dotbot.Gateway
-This web api uses Discord's interaction API functionality to drive a discord bot that uses slash commands
