--- src/Dotbot.Gateway/Settings/Discord.cs
@@ -1,7 +1,0 @@
-namespace Dotbot.Gateway.Settings;
-
-public class Discord
-{
-    public string Token { get; init; } = null!;
-    public string BucketEnvPrefix { get; set; } = null!;
-}
