+++ tests/Dotbot.Api.UnitTests/Dotbot.Api.UnitTests.csproj
@@ -1,0 +1,33 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <IsTestProject>true</IsTestProject>
+        <EnableMSTestRunner>true</EnableMSTestRunner>
+        <TestingPlatformDotnetTestSupport>true</TestingPlatformDotnetTestSupport>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Bogus" />
+        <PackageReference Include="BouncyCastle.Cryptography" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" />
+        <PackageReference Include="Microsoft.Testing.Extensions.TrxReport" />
+        <PackageReference Include="RichardSzalay.MockHttp" />
+        <PackageReference Include="TUnit" />
+
+        <PackageReference Include="NSubstitute"/>
+
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\src\Dotbot.Api\Dotbot.Api.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="res\" />
+      <Folder Include="TestResults\" />
+    </ItemGroup>
+    
+</Project>
