--- tests/Dotbot.Gateway.UnitTests/Dotbot.Gateway.UnitTests.csproj
@@ -1,33 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <IsTestProject>true</IsTestProject>
-        <EnableMSTestRunner>true</EnableMSTestRunner>
-        <TestingPlatformDotnetTestSupport>true</TestingPlatformDotnetTestSupport>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Bogus" />
-        <PackageReference Include="BouncyCastle.Cryptography" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" />
-        <PackageReference Include="Microsoft.Testing.Extensions.TrxReport" />
-        <PackageReference Include="RichardSzalay.MockHttp" />
-        <PackageReference Include="TUnit" />
-
-        <PackageReference Include="NSubstitute"/>
-
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\src\Dotbot.Gateway\Dotbot.Gateway.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="res\" />
-      <Folder Include="TestResults\" />
-    </ItemGroup>
-    
-</Project>
